+++ test/features/variable/src/api/functional/variable/$delete/index.ts
@@ -12,6 +12,35 @@ import type { IConnection } from "@nestia/fetcher";
  * @path DELETE /variable/delete
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
+export async function $$delete(
+    connection: IConnection,
+): Promise<void> {
+    return Fetcher.fetch(
+        connection,
+        $$delete.ENCRYPTED,
+        $$delete.METHOD,
+        $$delete.path(),
+    );
+}
+export namespace $$delete {
+
+    export const METHOD = "DELETE" as const;
+    export const PATH: string = "/variable/delete";
+    export const ENCRYPTED: Fetcher.IEncrypted = {
+        request: false,
+        response: false,
+    };
+
+    export const path = (): string => {
+        return `/variable/delete`;
+    }
+}
+
+/**
+ * @controller VariableController.$delete()
+ * @path GET /variable/delete
+ * @nestia Generated by Nestia - https://github.com/samchon/nestia
+ */
 export async function $delete(
     connection: IConnection,
 ): Promise<void> {
@@ -24,7 +53,7 @@ export async function $delete(
 }
 export namespace $delete {
 
-    export const METHOD = "DELETE" as const;
+    export const METHOD = "GET" as const;
     export const PATH: string = "/variable/delete";
     export const ENCRYPTED: Fetcher.IEncrypted = {
         request: false,
