--- test/features/variable/src/api/functional/variable/dollarDelete/index.ts
@@ -1,37 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional.variable.dollarDelete
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-import { Fetcher } from "@nestia/fetcher";
-import type { IConnection } from "@nestia/fetcher";
-
-/**
- * @controller VariableController.$delete()
- * @path DELETE /variable/dollarDelete
- * @nestia Generated by Nestia - https://github.com/samchon/nestia
- */
-export async function $delete(
-    connection: IConnection,
-): Promise<void> {
-    return Fetcher.fetch(
-        connection,
-        $delete.ENCRYPTED,
-        $delete.METHOD,
-        $delete.path(),
-    );
-}
-export namespace $delete {
-
-    export const METHOD = "DELETE" as const;
-    export const PATH: string = "/variable/dollarDelete";
-    export const ENCRYPTED: Fetcher.IEncrypted = {
-        request: false,
-        response: false,
-    };
-
-    export const path = (): string => {
-        return `/variable/dollarDelete`;
-    }
-}
