+++ test/features/variable/src/test/features/api/automated/test_api_variable_$delete_$$delete.ts
@@ -1,9 +1,9 @@
 import api from "./../../../../api";
 
-export const test_api_variable_dollarDelete_$delete = async (
+export const test_api_variable_$delete_$$delete = async (
     connection: api.IConnection
 ): Promise<void> => {
-    await api.functional.variable.dollarDelete.$delete(
+    await api.functional.variable.$delete.$$delete(
         connection,
     );
 };
